backend/requirements.txt
Flask
Flask-CORS
Flask-SocketIO 
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart 
python-dotenv

    # socketio and aioice are likely already installed given your app.py
    # If not, or to ensure they are, you can add:
    # python-socketio
    # aioice 